requirements.txt
boto3==1.28.44
chardet==5.2.0
embedchain==0.0.58
extra_streamlit_components==0.1.60
mailchimp3==3.0.21
openai==1.6.1
pdf2image==1.16.3
Pillow==9.5.0
PyJWT==2.8.0
pymongo==4.5.0
pypdfium2==4.20.0
pytesseract==0.3.10
python-dotenv==1.0.0
bcrypt==4.0.1
PyYAML==6.0.1
PyYAML==6.0.1
Requests==2.31.0
streamlit==1.26.0
stripe==6.5.0
